Refactor BurstTrie and add unit tests
Changed MaxContainerSize to float and made it adaptive on node burst. Fixed ContainerNode.GetAll to use AddRange. Improved InternalNode robustness in Insert and Search. Added BurstTrieTests for insertion, search, and retrieval validation.

+++ DataStructures/Trees/BurstTrie.cs
@@ -8,7 +8,7 @@ namespace DataStructures.Trees
 {
     public class BurstTrie
     {
-        public int MaxContainerSize;
+        public float MaxContainerSize;
         private BurstNode root;
         public int Count => root.Count;
         public BurstTrie(int maxContainerSize = 10)
